# Dual 1536/2048-bit exponentiation optimization for Intel IceLake CPU