@amatyuko-intc 2021-06-09
I have rebased the PR. Current 'master' branch has commits that rename functions related to RSA2k previous pull request (RSAZ -> ossl_rsaz), so I'll went ahead and extended this renaming to the current PR.