@paulidale 2021-06-09
Adding an OTC hold based on the closeness to beta criteria used to not merge #15361 and others.