@dtzimmerman 2021-06-10
Single Core/ Single Thread RSA 3K/4K Sign Performance on Intel(R) Xeon(R) Platinum 8368 CPU @ 2.40GHz (Turbo Disabled):

```
Baseline:
                  sign    verify     sign/s verify/s
rsa 3072 bits 0.002660s 0.000054s    376.0  18565.7
rsa 4096 bits 0.006177s 0.000094s    161.9  10678.0

With PR 14908:
                  sign    verify     sign/s verify/s
rsa 3072 bits 0.001202s 0.000054s    832.0  18551.6  --> 2.21x Sign Perf Improvement
rsa 4096 bits 0.002426s 0.000093s    412.2  10799.7  --> 2.55x Sign Perf Improvement
```