@richsalz 2021-06-10
I hope the OTC approves this.  This isn't quite the same as other PR's that were post-3.0, this is tuning code (with contributions from the CPU creator :) that gets a pretty amazing performance increase.